Simplified and Raw Coordinates
Overlays the MRCP image onto a plotly graph and calculates the fractal dimension of the overall contour. It simplifies the contour using the Ramer-Douglas-Peucker algorithm and visualizes both the original and simplified contours overlaid on the original image using Plotly. Additionally, it identifies and shades regions where the contour is not simplified enough, based on a ratio threshold, to aid further analysis.